First commit towards automated tests
I've created a single test for the java parser which should make it very easy
for us to create and add new test cases.

In automated-tests/resources/javaparser (we can renamed it to just 'test' when the old test
folder is removed) folder there are two subfolder: expected, source.

To test the output we simply create a .txt file for the java source and drop it in the source folder
and write the expected simple java and put it in a .txt file and drop it in the expected folder. The
test will then read and parse each file in source and compare the result with the contents of the
file in the expected folder with the same name.

So far I've just created a single file: SimpleClass to show that it does indeed work :)