Fixed bug with missing inferred 'this' field acc
The problem was that the lvars Set[String] that is used to keep
track of local variables didn't get reset when transforming different
classes so it would think that a local variables or argument specified
in a previous class would be accessible in the following classes and
therefore it wouldn't infer the 'this' field access.

Added test cases that shows it now works.

Note: We might need to change the algorithm of when 'this' should be
inferred. Currently we say that as long as the name isn't in the set
of local variables it most be a property but this won't work for
nested classes.