Fixed bug in purity analysis
Couldn't figure out why it didn't introduce any OutsideEdges so I re-read
the paper. Turns out that the way I found escaped nodes was very broken.

This is how I'm now finding escaped nodes: If a node is reachable from
any of the nodes in the following set then it is an escaped node:

  parameters of function ++ return value of funciton ++ globally escaped nodes

I use a BFS traversal taking each of the nodes in the above set as the root
and if any of the BFS traversals can reach the node in question then it is
escaped.