Didn't always find all the nodes in the graph when needed
There are times where I need to access all the nodes of the graph.
Previously I had just found all the nodes in the edges but that means
I had forgotten about the nodes that aren't connected by edges - I.E.
the nodes that are only stored in the StateOfLocalVariables map.