bug: 'this' didn't get mapped properly when combining graphs
Because the 'this' argument is implicit I had forgotten to map it
in the mapping that takes care of parameter nodes. I noticed it
because p.flip() which modified this.x & this.y didn't result
in any modifications to 'p' being recorded in the callers context.

Also made labels inside nodes more meaningful