Fixed while construct transfer function
After a lot of thought and a chat with Hannes I fixed a problem with
the transfer function for the while construct. An executive summary of
this problem:

There was no strategy for merging points-to graphs after each iteration
of a loop so if the loop body created a new Node (fx by storing a property
of an object in a local varible) a fixed point would never be reached. If
the loop body only invoked methods (such as the loop in flipAll) then this
issue didn't because the mapping for SJCalls took care of the nodes.