Fixed small issue in while loop merge [purity]
Fixed a small issue in the merging of the PT-Graphs when iterating
the while loop body. It's only okay to use an existing node if there's
an edge from the same object AND the same label that points to a new
object. Previously I didn't make sure that the label was the same