Added Playground objects to make it easier to interact with the code
Added a Playground.scala file that contains objects with main methods
for:

- parse a Java file and output the Java program and resulting
  non-optimized Simple Java program.
- parse a Java file and output the Java program, resulting non-
  optimized Simple Java program and optimized Simple Java program

The purpose of these objects is to make it easier to quickly check
how our different transformations of the program act.

Also added a removeDeadVariables to the Optimizer object that will
traverse an entire Java program and remove dead variables from all
of the methods.