rewrite to use lengthPositional instead of positional
This is supposed to include offset and length into the
SimpleJavaAST - supposedly using CharArrayReaderX and
a custom Scanner

Unfortunately it's a token stream where the position
offset is computed - thus we have token offsets rather
than byte offsets - which fails in practise

The resulting position information is pretty useless,
therefore committing onto a branch instead of master.

Most likely this will not be merged at all, but another
solution is desirable, which consists of correct offset
positions on the byte level to get the coloring in a
good shape.