Added an experimental Python implementation of the Coqoon build system
This isn't complete yet -- in particular, it doesn't actually support
project configurations other than the default one -- but it will eventually
replace the existing Makefile generation code and will become the way to
work on Coqoon projects without using Coqoon.

(As a historical note, I tried implementing this as a POSIX shell script at
first for the sake of portability, but -- surprise! -- it turns out the
POSIX shell is an utterly wretched programming environment. This Python
script, which took me about a day and a half of work, is much better than
the shell script was after almost three weeks...)