Simplified the code that selects build candidates
A file now becomes a build candidate if it can be compiled (that is, if all
of its dependencies can be compiled and yet don't need to be compiled) and
if it needs to be compiled (i.e., if it's older than one of its
dependencies). I'm not really sure what the old code did, but I don't think
it was this!