Use the new "-Q" option instead of "-I -as" when Coq 8.5 is detected
The check is actually the other way round: we use "-I -as" only when Coq
8.4 is detected, under the assumption that -Q is here to stay.

(As an aside, commit e33bc34be5ffd233294ec8766cee501c88622921 broke the
namespace enforcement option: a coqdir of "Foo.Bar" would be rendered as
the arguments "-I path/ -as Foo -as Bar". Oops! Luckily nobody noticed
until this commit fixed it...)