Reworked the way the Python build script handles the -Q option
"-Q dir/ Foo" is not, after all, equivalent to "-I dir/ -as Foo"; in fact,
it's essentially -R, but it requires names to be fully qualified. What's
worse, the special qualification behaviour overrides previous load path
entries for the same folders!

(This commit also makes a comment about load path resolution more helpful
and replicates it in both the Python build script and the Coqoon compiler
runner.)