Drastically improved the way the Python build script handles external paths
External paths used to be given a really unhelpful computed variable name
like "EXT_2", which was opaque, unhelpful, and vulnerable to collisions if
one project depended on another. Now /two/ predictable variable names are
used: one based on the original path, and another based on the coqdir (if
there is one). This turns the collision problem into a feature, and makes
the build script much better at realising when a dependency's already been
satisfied