Added support for more coqidetop commands and XML representations
About and Goal now join the gang. Whenever Coqoon gets a reply to an Add,
if Status indicates that a proof is in progress, Goal should be issued;
that'll cause the goal structure that the goal viewer needs to be emitted
at every relevant state, so the instant update from the PIDE editor will
still work. ... Maybe this will all work?